requirements.txt
Babel==1.3
Chameleon==2.14
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.18
PasteDeploy==1.5.2
pyramid==1.4.5
pyramid-jinja2==1.10
pyScss==1.2.0.post3
pytz==2013.9
PyYAML==3.10
repoze.lru==0.6
six==1.5.2
translationstring==1.1
venusian==1.0a8
WebOb==1.3.1
wsgiref==0.1.2
zope.deprecation==4.1.0
zope.interface==4.0.5
